extracted inspect_item_handler
